fixing deps, adding some logs, adding url request params

+++ main.py
@@ -1,7 +1,9 @@
+import os
 import time
+import urllib.parse
 import urllib.request
 
-from flask import Flask
+from flask import Flask, request
 
 from mms.align_utils import (
     DEVICE,
@@ -17,39 +19,57 @@ app = Flask(__name__)
 
 @app.route("/")
 def hello_world():
-    # TODO: The following should be passed in the request.
-    wav_url = (
-        "https://firebasestorage.googleapis.com/v0/b/waha-ai-timestamper-4265a.appspot.com/"
-        "o/sessions%2F1fdbc3e9-6aab-4beb-9277-c0e1390ae345%2Ftest%20copy.wav?alt=media"
-    )
-    txt_url = (
-        "https://firebasestorage.googleapis.com/v0/b/waha-ai-timestamper-4265a.appspot.com/"
-        "o/sessions%2F1fdbc3e9-6aab-4beb-9277-c0e1390ae345%2Ftest%20copy.txt?alt=media"
-    )
-    language = "eng"
+    language = request.args.get("lang")
+    audio_file = request.args.get("audio-file")
+    text_file = request.args.get("text-file")
+    session_id = request.args.get("session-id")
 
-    wav_file = "test.wav"
-    txt_file = "test.txt"
+    if audio_file is None:
+        return "Missing audio-file parameter", 400
+    elif text_file is None:
+        return "Missing text-file parameter", 400
+    elif language is None:
+        return "Missing lang parameter", 400
+    elif session_id is None:
+        return "Missing session-id parameter", 400
 
-    urllib.request.urlretrieve(
-        wav_url,
-        wav_file,
-    )
-    urllib.request.urlretrieve(
-        txt_url,
-        txt_file,
-    )
-    lines_to_timestamp = open(txt_file, "r").read().split("\n")
+    base_url = "https://firebasestorage.googleapis.com/v0/b/waha-ai-timestamper-4265a.appspot.com/o/"
+    audio_path = f"sessions/{session_id}/{audio_file}"
+    text_path = f"sessions/{session_id}/{text_file}"
+    audio_path_encoded = urllib.parse.quote(audio_path, safe="")
+    text_path_encoded = urllib.parse.quote(text_path, safe="")
+    audio_url = f"{base_url}{audio_path_encoded}?alt=media"
+    text_url = f"{base_url}{text_path_encoded}?alt=media"
+    print(audio_url, text_url)
+
+    if not os.path.exists(audio_file):
+        print("Downloading audio... ")
+        urllib.request.urlretrieve(
+            audio_url,
+            audio_file,
+        )
+
+    if not os.path.exists(text_file):
+        print("Downloading text... ")
+        urllib.request.urlretrieve(
+            text_url,
+            text_file,
+        )
+
+    print("Normalizing and romanizing... ")
+    lines_to_timestamp = open(text_file, "r", encoding="utf-8").read().split("\n")
     norm_lines_to_timestamp = [
         text_normalize(line.strip(), language) for line in lines_to_timestamp
     ]
     uroman_lines_to_timestamp = get_uroman_tokens(norm_lines_to_timestamp, language)
 
+    print("Loading model.... ")
     model, dictionary = get_model_and_dict()
     model = model.to(DEVICE)
 
+    print("Aligning...")
     segments, stride = get_alignments(
-        wav_file,
+        audio_file,
         uroman_lines_to_timestamp,
         model,
         dictionary,
